extract tokenizer to separate class
