--- src/Highlighter/Token/CommentToken.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace VStelmakh\UrlHighlight\Highlighter\Token;
-
-/**
- * @internal
- */
-final readonly class CommentToken extends AbstractToken {}
