--- src/Highlighter/Token/PlainToken.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace VStelmakh\UrlHighlight\Highlighter\Token;
-
-/**
- * @internal
- */
-final readonly class PlainToken extends AbstractToken {}
