+++ src/Highlighter/Tokenizer/Token/Token.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace VStelmakh\UrlHighlight\Highlighter\Tokenizer\Token;
+
+/**
+ * @internal
+ */
+interface Token
+{
+    public function toString(): string;
+}
